fix ts test
